最终整合7/16

+++ src/main/java/com/neusoft/po/Lesson.java
@@ -6,7 +6,7 @@ public class Lesson {
 	private String lname;			//课程名
 
 	private String imgUrl;			//图片url
-	private float lprice;			//课程价格
+	private double lprice;			//课程价格
 	private String ldesc;			//课程富文本
 	private String category;		//课程类别
 	private String pubTime;			//发布时间
@@ -32,10 +32,10 @@ public class Lesson {
 	public void setImgUrl(String imgUrl) {
 		this.imgUrl = imgUrl;
 	}
-	public float getLprice() {
+	public double getLprice() {
 		return lprice;
 	}
-	public void setLprice(float lprice) {
+	public void setLprice(double lprice) {
 		this.lprice = lprice;
 	}
 	public String getLdesc() {
